+++ src/FasTnT.Host/Controllers/v1_2/EpcisXmlQueryController.cs
@@ -6,11 +6,11 @@ using Microsoft.AspNetCore.Mvc;
 namespace FasTnT.Host.Controllers.v1_2
 {
     [Authorize]
-    [Formatter(Format.Soap)]
-    [Route("v1_2/Query.svc")]
-    public class EpcisSoapQueryService : EpcisQueryController
+    [Formatter(Format.Xml)]
+    [Route("v1_2/Query")]
+    public class EpcisXmlQueryController : EpcisQueryController
     {
-        public EpcisSoapQueryService(QueryDispatcher dispatcher) : base(dispatcher)
+        public EpcisXmlQueryController(QueryDispatcher dispatcher) : base(dispatcher)
         {
         }
     }
